Class User created, missing assing values from dictionary and unittests
