+++ models/users.py~
@@ -1,0 +1,21 @@
+#!/usr/bin/python3
+""" holds class User"""
+from models.baseModel import BaseModel, Base
+from sqlalchemy import Column, String
+
+
+class User(BaseModel, Base):
+    """Representation of a user """
+    __tablename__ = 'Users'
+    id = Column(String(40), nullable=False, primary_key=True)
+    FirstName = Column(String(50), nullable=False)
+    LastName = Column(String(50), nullable=False)
+    Phone = Column(String(30), nullable=False)
+    Mail = Column(String(50), nullable=True)
+    Password = Column(String(50), nullable=False)
+    City = Column(String(50), nullable=False)
+
+    def __init__(self, *args):
+        """initializes user"""
+        super().__init__(*args)
+
